+++ src/app.vue
@@ -2,7 +2,7 @@
 	<error></error>
 	<update></update>
 	<settings></settings>
-	<spotify_player></spotify_player>
+	<youtube_player></youtube_player>
 </template>
 
 <script setup lang="ts">
@@ -10,7 +10,7 @@ console.vLog("Initializing vue");
 import error from "./components/error.vue";
 import update from "./components/update.vue";
 import settings from "./components/settings.vue";
-import spotify_player from "./components/spotifyplayer.vue";
+import youtube_player from "./components/youtubeplayer.vue";
 
 console.vLog("Initialized!");
 </script>